+++ leptos-fluent/src/lib.rs
@@ -14,7 +14,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! leptos-fluent = "0.0.31"
+//! leptos-fluent = "0.0.32"
 //! fluent-templates = "0.9"
 //!
 //! [features]