+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased - [0.0.32]
+## 2024-06-03 - [0.0.32]
 
 ### Breaking changes
 
@@ -167,7 +167,7 @@
 
 - Added all ISO-639-1 and ISO-639-2 languages.
 
-[0.0.32]: https://github.com/mondeja/leptos-fluent/compare/v0.0.31...master
+[0.0.32]: https://github.com/mondeja/leptos-fluent/compare/v0.0.31...v0.0.32
 [0.0.31]: https://github.com/mondeja/leptos-fluent/compare/v0.0.30...v0.0.31
 [0.0.30]: https://github.com/mondeja/leptos-fluent/compare/v0.0.29...v0.0.30
 [0.0.29]: https://github.com/mondeja/leptos-fluent/compare/v0.0.28...v0.0.29